+++ var/spack/repos/requirements.test/repo.yaml
@@ -1,0 +1,2 @@
+repo:
+  namespace: requirements.test
